+++ src/main/java/frc/robot/RobotContainer.java
@@ -33,22 +33,22 @@ import frc.robot.commands.TriggerEventCmd;
 import frc.robot.commands.WaitForTowerStateCmd;
 import frc.robot.commands.WaitToSeeCoralCmd;
 import frc.robot.generated.TunerConstants;
-import frc.robot.subsystems.ApproachSubsystem;
-import frc.robot.subsystems.ApproachTarget;
-import frc.robot.subsystems.ElevatorSubsystem;
 import frc.robot.subsystems.Globals;
-import frc.robot.subsystems.LEDSubsystem;
-import frc.robot.subsystems.TowerEvent;
-import frc.robot.subsystems.TowerState;
-import frc.robot.subsystems.TowerSubsystem;
-import frc.robot.subsystems.VisionSubsystem;
-import frc.robot.subsystems.WristSubsystem;
+import frc.robot.subsystems.approach.ApproachSubsystem;
+import frc.robot.subsystems.approach.ApproachTarget;
 import frc.robot.subsystems.drive.Drive;
 import frc.robot.subsystems.drive.GyroIO;
 import frc.robot.subsystems.drive.GyroIOPigeon2;
 import frc.robot.subsystems.drive.ModuleIO;
 import frc.robot.subsystems.drive.ModuleIOSim;
 import frc.robot.subsystems.drive.ModuleIOTalonFX;
+import frc.robot.subsystems.elevator.ElevatorSubsystem;
+import frc.robot.subsystems.led.LEDSubsystem;
+import frc.robot.subsystems.tower.TowerEvent;
+import frc.robot.subsystems.tower.TowerState;
+import frc.robot.subsystems.tower.TowerSubsystem;
+import frc.robot.subsystems.vision.VisionSubsystem;
+import frc.robot.subsystems.wrist.WristSubsystem;
 
 public class RobotContainer {
     private final CommandXboxController joystick = new CommandXboxController(0);
