+++ src/main/java/frc/robot/commands/JustIntakeCmd.java
@@ -5,9 +5,9 @@
 package frc.robot.commands;
 
 import edu.wpi.first.wpilibj2.command.Command;
-import frc.robot.subsystems.TowerEvent;
-import frc.robot.subsystems.TowerState;
-import frc.robot.subsystems.TowerSubsystem;
+import frc.robot.subsystems.tower.TowerEvent;
+import frc.robot.subsystems.tower.TowerState;
+import frc.robot.subsystems.tower.TowerSubsystem;
 
 public class JustIntakeCmd extends Command {
   TowerSubsystem tower;
