+++ src/main/java/frc/robot/commands/TriggerEventCmd.java
@@ -5,8 +5,8 @@
 package frc.robot.commands;
 
 import edu.wpi.first.wpilibj2.command.Command;
-import frc.robot.subsystems.TowerEvent;
-import frc.robot.subsystems.TowerSubsystem;
+import frc.robot.subsystems.tower.TowerEvent;
+import frc.robot.subsystems.tower.TowerSubsystem;
 
 public class TriggerEventCmd extends Command {
   TowerSubsystem tower;
