+++ src/main/java/frc/robot/commands/WaitToSeeCoralCmd.java
@@ -5,8 +5,8 @@
 package frc.robot.commands;
 
 import edu.wpi.first.wpilibj2.command.Command;
-import frc.robot.subsystems.TowerState;
-import frc.robot.subsystems.TowerSubsystem;
+import frc.robot.subsystems.tower.TowerState;
+import frc.robot.subsystems.tower.TowerSubsystem;
 
 public class WaitToSeeCoralCmd extends Command {
   TowerSubsystem tower;
