+++ src/main/java/frc/robot/subsystems/Globals.java
@@ -2,6 +2,8 @@ package frc.robot.subsystems;
 
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
+import frc.robot.subsystems.approach.ApproachTarget;
+import frc.robot.subsystems.led.LEDmode;
 
 public final class Globals extends SubsystemBase {
