+++ src/main/java/frc/robot/subsystems/approach/ApproachTarget.java
@@ -2,7 +2,7 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.subsystems;
+package frc.robot.subsystems.approach;
 
 import com.pathplanner.lib.path.GoalEndState;
 import edu.wpi.first.math.geometry.Pose2d;
