+++ src/main/java/frc/robot/subsystems/elevator/ElevatorSubsystem.java
@@ -2,7 +2,7 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.subsystems;
+package frc.robot.subsystems.elevator;
 
 import com.revrobotics.spark.SparkBase.ControlType;
 import com.revrobotics.spark.SparkBase.PersistMode;
@@ -27,6 +27,7 @@ import edu.wpi.first.wpilibj2.command.SubsystemBase;
 import frc.robot.Constants;
 import frc.robot.Constants.Elevator;
 import frc.robot.commands.DefaultElevatorCmd;
+import frc.robot.subsystems.Globals;
 
 public class ElevatorSubsystem extends SubsystemBase {
   private final SparkFlex leftElevatorMotor;
