+++ src/main/java/frc/robot/subsystems/led/LEDmode.java
@@ -1,4 +1,4 @@
-package frc.robot.subsystems;
+package frc.robot.subsystems.led;
 
 public enum LEDmode {
     NONE,
