+++ src/main/java/frc/robot/subsystems/tower/TowerState.java
@@ -1,4 +1,4 @@
-package frc.robot.subsystems;
+package frc.robot.subsystems.tower;
 
 public enum TowerState {
 	//Start
