+++ src/main/java/frc/robot/subsystems/vision/VisionSubsystem.java
@@ -1,5 +1,5 @@
 
-package frc.robot.subsystems;
+package frc.robot.subsystems.vision;
 
 import java.util.Optional;
